+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `database::status64` wrapping `sqlite3_db_status64`
 - `database::setlk_timeout` wrapping `sqlite3_setlk_timeout`
 - `database::set_errmsg` wrapping `sqlite3_set_errmsg`
+- `keywords` class providing safe, STL-like interface to `sqlite3_keyword_count`, `sqlite3_keyword_name` and `sqlite3_keyword_check`
 - clang-cl is now supported on Windows
 
 ### Fixed